+++ changelog.md
@@ -1,8 +1,12 @@
 # The Prestige Tree: Rewritten changelog (major updates only):
 
+### v0.3: Enhanced Spacetime - 10/30/20
+- Ported the "T", "E", & "S" layers
+- Added the next group of achievements
+
 ### v0.2: The Boosted Gen - 10/29/20
 - Ported the "B" & "G" layers
-- Added the next X achievements
+- Added the next group of achievements
 
 ### v0.1: Quite Prestigious - 10/28/20
 - Ported the "P" layer